+++ src/utils/objloader.cpp
@@ -452,7 +452,7 @@ bool ObjLoader::LoadMaterial(const string& filename, Material* const mat)
          float tmp;
          fscanf_s(f, "%f\n", &tmp);
          const int d = (int)(tmp * 100.f + 0.5f);
-         tmp = d / 100.0f;
+         tmp = (float)d / 100.0f;
          // normally a wavefront material specular exponent ranges from 0..1000.
          // but our shininess calculation differs from the way how e.g. Blender is calculating the specular exponent
          // starting from 0.5 and use only half of the exponent resolution to get a similar look
